move time functions to utils
