+++ front/src/components/form/DayTimeSlider.jsx
@@ -5,8 +5,7 @@ import InputRange from 'react-input-range';
 import './PeriodRangeSlider.scss'; //TODO
 
 import Question from '../../Question';
-
-const DaySec = 86400; // 24 hours
+import { prettyHours, DaySeconds } from '../../Utils';
 
 // bg gradient
 const wrapperStyle = {};
@@ -26,23 +25,6 @@ const cellStyle = function(percent) {
         textAlign,
     };
 };
-
-const prettyHours = function(sec) {
-    let hours   = Math.floor(sec / 3600);
-    let minutes = Math.floor((sec - (hours * 3600)) / 60);
-    let seconds = sec - (hours * 3600) - (minutes * 60);
-
-    if (hours   < 10) {hours   = "0" + hours;}
-    let t = 'am';
-    if (hours > 12) {
-        hours -= 12;
-        t = 'pm';
-    }
-    if (minutes < 10) {minutes = "0" + minutes;}
-    if (seconds < 10) {seconds = "0" + seconds;}
-    return `${hours}:${minutes} ${t}`;
-};
-
 
 /**
  * range sliders for defining period secondss (seconds) within 24 hours
@@ -67,18 +49,21 @@ class DayTimeSlider extends Component {
 
     render() {
         const { value } = this.state;
-
+        const { withIcons} = this.props;
         return (
             <div style={ wrapperStyle }>
-                <div className={ this.props.className } style={ tableStyle }>
-                    <div style={ cellStyle(0) }><i className="fas fa-moon"></i></div>
-                    <div style={ cellStyle(50) }><i className="fas fa-sun"></i></div>
-                    <div style={ cellStyle(100) }><i className="fas fa-moon"></i></div>
-                </div>
+                {
+                    withIcons &&
+                    <div className={ this.props.className } style={ tableStyle }>
+                        <div style={ cellStyle(0) }><i className="fas fa-moon"></i></div>
+                        <div style={ cellStyle(50) }><i className="fas fa-sun"></i></div>
+                        <div style={ cellStyle(100) }><i className="fas fa-moon"></i></div>
+                    </div>
+                }
                 <InputRange
                     className=""
                     minValue={ 0 }
-                    maxValue={ DaySec }
+                    maxValue={ DaySeconds }
                     value={ value }
                     onChange={ this.handleChange.bind(this) }
                     formatLabel={ val => prettyHours(val) }
@@ -96,6 +81,7 @@ DayTimeSlider.defaultProps = {
 
     // react-input-range props
     step:  15 * 60,
+    withIcons: true,
 };
 
 DayTimeSlider.propTypes = {
@@ -104,6 +90,7 @@ DayTimeSlider.propTypes = {
     required: PropTypes.bool,
     timeBeginLabel: PropTypes.string,
     timeEndLabel: PropTypes.string,
+    withIcons: PropTypes.bool,
 
     // react-input-range props
     step: PropTypes.number,
